move properties helper into events file
